+++ plugins/catalog-backend/src/schema/openapi.generated.ts
@@ -191,12 +191,69 @@ export const spec = {
       },
     },
     requestBodies: {},
-    responses: {},
+    responses: {
+      ErrorResponse: {
+        description: 'An error response from the backend.',
+        content: {
+          'application/json': {
+            schema: {
+              $ref: '#/components/schemas/Error',
+            },
+          },
+        },
+      },
+    },
     schemas: {
+      Error: {
+        type: 'object',
+        properties: {
+          error: {
+            type: 'object',
+            properties: {
+              name: {
+                type: 'string',
+              },
+              message: {
+                type: 'string',
+              },
+              stack: {
+                type: 'string',
+              },
+              code: {
+                type: 'string',
+              },
+            },
+            required: ['name', 'message'],
+          },
+          request: {
+            type: 'object',
+            properties: {
+              method: {
+                type: 'string',
+              },
+              url: {
+                type: 'string',
+              },
+            },
+            required: ['method', 'url'],
+          },
+          response: {
+            type: 'object',
+            properties: {
+              statusCode: {
+                type: 'number',
+              },
+            },
+            required: ['statusCode'],
+          },
+        },
+        required: ['error', 'response'],
+      },
       JsonObject: {
         type: 'object',
         properties: {},
         description: 'A type representing all allowed JSON object values.',
+        additionalProperties: true,
       },
       MapStringString: {
         type: 'object',
@@ -698,38 +755,6 @@ export const spec = {
         required: ['type', 'target'],
         additionalProperties: false,
       },
-      SerializedError: {
-        allOf: [
-          {
-            $ref: '#/components/schemas/JsonObject',
-          },
-          {
-            type: 'object',
-            properties: {
-              code: {
-                type: 'string',
-                description:
-                  'A custom code (not necessarily the same as an HTTP response code); may not be present',
-              },
-              stack: {
-                type: 'string',
-                description: 'A stringified stack trace; may not be present',
-              },
-              message: {
-                type: 'string',
-                description: 'The message of the exception that was thrown',
-              },
-              name: {
-                type: 'string',
-                description: 'The name of the exception that was thrown',
-              },
-            },
-            required: ['message', 'name'],
-          },
-        ],
-        description: 'The serialized form of an Error.',
-        additionalProperties: false,
-      },
       EntitiesQueryResponse: {
         type: 'object',
         properties: {
@@ -776,6 +801,12 @@ export const spec = {
         responses: {
           '200': {
             description: 'Refreshed',
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -828,6 +859,12 @@ export const spec = {
                 },
               },
             },
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -882,6 +919,12 @@ export const spec = {
               },
             },
           },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
+          },
         },
         security: [
           {},
@@ -901,6 +944,12 @@ export const spec = {
         responses: {
           '204': {
             description: 'Deleted successfully.',
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -930,6 +979,12 @@ export const spec = {
                 },
               },
             },
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -965,6 +1020,12 @@ export const spec = {
                 },
               },
             },
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -1002,6 +1063,12 @@ export const spec = {
               },
             },
           },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
+          },
         },
         security: [
           {},
@@ -1015,7 +1082,7 @@ export const spec = {
           },
         ],
         requestBody: {
-          required: true,
+          required: false,
           content: {
             'application/json': {
               schema: {
@@ -1071,6 +1138,12 @@ export const spec = {
                 },
               },
             },
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -1139,6 +1212,12 @@ export const spec = {
               },
             },
           },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
+          },
         },
         security: [
           {},
@@ -1203,6 +1282,44 @@ export const spec = {
               },
             },
           },
+          '201': {
+            description: '201 response',
+            content: {
+              'application/json': {
+                schema: {
+                  type: 'object',
+                  properties: {
+                    location: {
+                      type: 'object',
+                      properties: {
+                        id: {
+                          type: 'string',
+                        },
+                        type: {
+                          type: 'string',
+                        },
+                        target: {
+                          type: 'string',
+                        },
+                      },
+                      required: ['id', 'type', 'target'],
+                    },
+                    entities: {
+                      type: 'array',
+                      items: {},
+                    },
+                  },
+                  required: ['location', 'entities'],
+                },
+              },
+            },
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
+          },
         },
         security: [
           {},
@@ -1264,6 +1381,9 @@ export const spec = {
               },
             },
           },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
+          },
         },
         security: [
           {},
@@ -1288,6 +1408,9 @@ export const spec = {
                 },
               },
             },
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -1314,6 +1437,12 @@ export const spec = {
         responses: {
           '204': {
             description: 'No content',
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -1349,6 +1478,12 @@ export const spec = {
                 },
               },
             },
+          },
+          '400': {
+            $ref: '#/components/responses/ErrorResponse',
+          },
+          default: {
+            $ref: '#/components/responses/ErrorResponse',
           },
         },
         security: [
@@ -1387,32 +1522,31 @@ export const spec = {
         responses: {
           '200': {
             description: 'Ok',
+          },
+          '400': {
+            description: '400 response',
             content: {
-              'application/json': {
+              'application/json; charset=utf-8': {
                 schema: {
-                  anyOf: [
-                    {
-                      type: 'object',
-                      properties: {
-                        errors: {
-                          $ref: '#/components/schemas/SerializedError',
-                        },
-                      },
-                      required: ['errors'],
-                    },
-                    {
-                      type: 'object',
-                      properties: {
-                        errors: {
-                          type: 'array',
-                          items: {
-                            $ref: '#/components/schemas/SerializedError',
+                  type: 'object',
+                  properties: {
+                    errors: {
+                      type: 'array',
+                      items: {
+                        type: 'object',
+                        properties: {
+                          name: {
+                            type: 'string',
+                          },
+                          message: {
+                            type: 'string',
                           },
                         },
+                        required: ['name', 'message'],
                       },
-                      required: ['errors'],
                     },
-                  ],
+                  },
+                  required: ['errors'],
                 },
               },
             },