+++ packages/app-defaults/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/app-defaults
 
+## 1.4.5-next.0
+
+### Patch Changes
+
+- 6c2b872153: Add official support for React 18.
+- Updated dependencies
+  - @backstage/core-components@0.13.7-next.0
+  - @backstage/core-plugin-api@1.8.0-next.0
+  - @backstage/plugin-permission-react@0.4.17-next.0
+  - @backstage/core-app-api@1.11.1-next.0
+  - @backstage/theme@0.4.4-next.0
+
 ## 1.4.4
 
 ### Patch Changes