+++ plugins/analytics-module-newrelic-browser/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-analytics-module-newrelic-browser
 
+## 0.0.4-next.0
+
+### Patch Changes
+
+- 6c2b872153: Add official support for React 18.
+- Updated dependencies
+  - @backstage/core-components@0.13.7-next.0
+  - @backstage/core-plugin-api@1.8.0-next.0
+  - @backstage/config@1.1.1
+
 ## 0.0.3
 
 ### Patch Changes