+++ plugins/code-coverage-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-code-coverage-backend
 
+## 0.2.21-next.0
+
+### Patch Changes
+
+- 11f671eaa9: Added support for new backend system
+- Updated dependencies
+  - @backstage/backend-common@0.19.9-next.0
+  - @backstage/integration@1.7.1
+  - @backstage/backend-plugin-api@0.6.7-next.0
+  - @backstage/catalog-client@1.4.5
+  - @backstage/catalog-model@1.4.3
+  - @backstage/config@1.1.1
+  - @backstage/errors@1.2.3
+  - @backstage/plugin-auth-node@0.4.1-next.0
+
 ## 0.2.20
 
 ### Patch Changes