+++ plugins/nomad-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-nomad-backend
 
+## 0.1.9-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.9-next.0
+  - @backstage/backend-plugin-api@0.6.7-next.0
+  - @backstage/config@1.1.1
+  - @backstage/errors@1.2.3
+
 ## 0.1.8
 
 ### Patch Changes